requirements.txt
streamlit>=1.51.0
python-dotenv>=1.0.1
langchain-ollama>=1.0.1
diffusers>=0.35.0
transformers>=4.51.0
accelerate>=1.6.0
safetensors>=0.4.3
torch>=2.8.0
Pillow>=11.3.0
